PsychVulkanCore: Fix mismatch cmdbuffer index vs. swapchain image index on macOS.
Make sure that when presenting a new image, we choose the right
command buffer for the currently acquired target swapchain image
that is about to be presented. We have a dedicated cmd buffer for
each swapchain image for copying the interop image into the target
swapchain image. If we choose the wrong cmd buffer for the swapchain
image, then stale content will be presented instead of the updated
content. The target index is window->currentSwapChainBuffer, not
window->frameIndex % numBuffers!

This worked nicely on well behaved operating systems and drivers like
Linux and Windows, where for a double-buffered swapchain setup the
sequence was always 0,1,0,1,0,1..... On macOS we get arbitrary indices
for target swapchain images, e.g., the same index on multiple successive
presents! While this itself looks like it could cause performance issues,
it should not cause false visual results, so use window->currentSwapChainBuffer
to prevent getting this wrong on weird systems like macOS + MoltenVK.

-> This fixes the weird judder and "images from the past" effect observed
with OpenGL+Vulkan interop, which turns out to be not an interop problem
at all, but just presentation of stale images due to use of wrong cmd buffer.

=> All our Vulkan demos now render and display correctly, as tested with
VBLSyncTest, PerceptualVBLSyncTest, LinesDemo and
AdditiveBlendingForLinearSuperpositionTutorial so far.

=> Visually correct and usable from here on. However, actual presentation
timing accuracy is not yet verified via external measurement equipment, so
this might still be wrong timing-wise. Also, performance is still disappointing,
with update rate capped at 1/2 video refresh rate, e.g., max 30 fps on the
60 Hz Retina display of the MBP 2017. None of the hacks, like setting the
CAMetalLayer as opaque, or disabling vsync on it, helped in any way.

=> New baseline, visually correct, timing unknown, performance disappointing.