PsychVulkanCore: Enable macOS direct-to-display mode for 10 bpc fb.
Our current code already enabled Metal Direct-To-Display mode for standard
8 bpc framebuffers under Vulkan + MoltenVK + Metal. Now we also enable it
for 10 bpc framebuffers.

For that we need to make sure that we create our swapchain images in the
exact 10 bpc format that the display engines are configured to use for 10 bit
scanout, which is the exact opposite (Apple!) of what all other operating
systems and Vulkan drivers use on AMD, NVidia and Intel for Linux and Windows.

I tested the availability of D-To-D mode by tracing with the Apple macOS
instruments profiling tools and the Metal Performance Trace profile, so i can
see what happens when wrt. graphics and display. The tool marks display
surfaces as fullscreen + matlab + DToD when they are DToD, so we can be
reasonably certain.

DToD for our Vulkan fullscreen, opaque, unoccluded, decorationless windows
under MoltenVK + Metal on macOS works at least as tested under:

- 8 bpc and 10 bpc Vulkan framebuffer mode, but not with fp16 mode.
- For both 8 bpc and 10 bpc display settings, ie. true scanout format.
- With the standard VK_COLOR_SPACE_SRGB_NONLINEAR_KHR color space,
but also tested with the VK_COLOR_SPACE_PASSTHROUGH_EXT space.
- Under native Retina panel resolution and scaled modes, at least after Screen()
did our usual Retina display setup. Iow. it works in collaboration with Screen.

Obviously any windowed or transparent mode would always trigger desktop
composition.

So in theory we got our fullscreen onscreen Vulkan windows configured by
the textbook, exactly in the way Apple recommends to get DToD mode for
maximum performance and minimal latency, bypassing the desktop compositor.

In practice, while the DC is mostly not involved (no gpu activity for compositing,
only a short invocation of about 10 microseconds to queue the DToD display
surface into the gpu display engine at next vblank, ie. flip programming), it is
not completely out of the picture. It seems to be needed for programming the
flip. And much worse: The compositor seems to get scheduled about 1-2 msecs
after the vblank which starts the next refresh cycle to do the hw programming,
instead of immediately at the Metal drawable present: request!!! It schedules
as if DToD is *not* active. This is different from OpenGL where the DC seems
to get activated at each swapbuffers, but it gets triggered immediately in
response to the CGLFlushDrawable request, ie. without adding significant
latency.

Under Vulkan, the beyond dumb scheduling of the DC in DToD mode adds
and enforced extra video frame of latency to any present request. While
this does not matter at all if one want to Screen('Flip') only every 2nd, 3rd, ...
frame, it is a complete performance killer for flipping at each refresh cycle,
as each Flip will take at least 2 refresh cycles to complete, cutting achievable
maximum redraw/present rate to half the video refresh rate of the display!

This as tested on a 2017 MacBookPro with AMD Polaris11 gpu under macOS
Catalina 10.15.7 with all updates and bug fixes as of 18th March 2021 and
MoltenVK 1.1.1.

E.g., on the MBP Retina panel with 60 Hz refresh i can get at most 30 fps.
On the 144 Hz external DP HDR monitor, i get at most 72 fps.

After spending over one day tracing all kinds of configurations, i'm pretty
sure that this is not a bug or flaw in PTB's Vulkan code, or the MoltenVK
Vulkan driver, but a severe flaw in macOS Metal or the WindowServer, with
no way for us to fix this or work around it.

One can disable all timestamping and run Flip completely blind and non-
synchronized to visual stimulus onset, and by avoiding any waiting on
flip completion one can avoid throttling to half video refresh rate. This means
there is still 1 frame of latency and chaotic unreliable timing, but one can
get full video refresh rate this way, in the very few scenarios where this would
make any sense for visual stimulation in a vision science context.

It is what it is. Up next: Checking if the reported timestamps have any relation
to reality by verifying with my external timing measurement equipment. Maybe
at least that will work if this is our lucky day...