PsychVulkanCore/OSX: Implement flag 0x2 as "disable DirectToDisplay".
Passing OpenWindow flag 0x2 will setup a swapchain format that prevents
use of Metal "Direct To Display" optimizations. For this, in 10 bpc mode, we
intentionally choose the RGB10A2 ordered format which is subject to
compositing, instead of the DToD optimized BGR10A2. In 8 bpc mode, we
fall back to RGB10A2, preventing DToD, instead of using the DToD BGRA8
format. Here however, we still choose a 8 bpc interop image format, to avoid
performance hits in Screen's imaging pipeline due to buffer unsharing.

Why would we want to intentionally allow setup of a suboptimal mode that
goes through the compositor? The mystery shall be solved in the next commit.