FlipTimingWithRTBoxPhotoDiodeTest: Do not ever use CoreVideo timestamping on macOS.
Retesting shows it is just as utterly broken on 10.15 as it was 15'ish years ago
on 10.4 Tiger. Sad. We should never enable it even as option. Probably i should
just rip that code out, i don't think the iToys company will fix this ever.
Currently it reports timestamps from roughly the middle of a scanout cycle...