PsychVulkan/OSX: Add special present path for macOS to work around its bugs.
Instead of hooking a present request to fast-path cmd 0, like on Linux and
Windows, we hook it to new cmd 2 on macOS. This implements the same
PsychVulkanCore present request, but does post-processing of timestamps
differently, to try to work around the deficiencies of the Apple OS:

1. If Vulkan delivers reasonable timestamp, check it against a vblank ts
computed by our regular PsychtoolboxkernelDriver method from Screen,
and report all larger than 1 msecs deviations between the two.

2. If Vulkan delivers invalid zero timestamps, fallback to the vblank ts, or
if that is unavailable to a simple GetSecs timestamp. Report this to user.

3. If Vulkan delivers the -1 present error timestamp, give up and print the
error message.

Testing shows that when Vulkan - actually macOS Metal - delivers a ts, then
it is roughly correct. Seems to be a vblank interrupt timestamp that serves as
proxy for a proper onset timestamp, so it is worse than what we do with our
own Screen + PTB kernel driver, but good enough for many use cases.
We see a couple of 100 usecs deviation.

If zero timestamps are returned, we can fill in with our vblank or GetSecs
timestamps, if those are available. Ofc. the whole point of this Vulkan on
macOS exercise is to not need the kernel driver et al. anymore, so in most
cases when we need this presentation backend, we won't have PTBK as
fallback. At least it is good enough for driver testing on my machine...

A -1 failure timestamp is unrecoverable - i tried to make it work, i failed.

-> So the verdict so far is that the macOS Metal timing/timestamping is just
another bug-fest, but it is not totally useless, just to a large degree...
Maybe retesting with more recent MoltenVK or similar will bring some
improvement, although i doubt it.