HDRViewer: Clamp mouse positions to valid range, handle out-of-window.
Screen('GetImage') does not like screenshot rects outside of the window,
and the rest of the code doesn't like empty samples from empty rects,
so deal with a mouse cursor positioned outside the HDR window, e.g., on
multi-display setups, or at the window borders.

Avoids error abort if mouse pointer gets moved to window borders or out
of window.