PsychVullkanCore/PsychHDR: Enable minimal HDR support for OSX.
On macOS 10.15, we use the same approach as on MS-Windows in windowed
mode. We select a scRGB VK_COLOR_SPACE_EXTENDED_SRGB_LINEAR_EXT
colorspace, which will enable HDR mode on macOS, with a linear encoding of
color values in a sRGB extended colorspace. This is supported on macOS
older than macOS 11, so actually usable and testable by myself, as opposed
to the standard HDR-10 + PQ color space, which is only supported starting
macOS 11. Theoretically also on 10.15, but due to Apple macOS SDK bugs
difficult to use, so the MoltenVK developers decided to restrict it to macOS 11+.

Anyway, we use scRGB linear encoding, with. a twist: As of MoltenVK 1.1.3,
their Metal CAEDRMetadata opticalOutputScale parameter is hard-coded to
1.0 in their vkSetHDRMetadataEXT() implementation, which means that pixel
values are supposed to be in units of nits, instead of units of multiple of 80 nits,
as the scRGB HDR standard demands. So we need to use a linear scaling factor
of 10000 instead of 125 to map the normalized [0;1] range to [0;10000] nits.

This seems to give reasonable results as far as light testing shows.
Note that macOS does not report the displays HDR properties, so our queries
of HDR display properties just return invalid "all zeros". This is a severe limitation
compared to proper operating systems like Linux and Windows in the HDR
department, but it is what it is...

-> HDR support for macOS is to be understood as proof-of-concept or alpha
quality at best. I do not guarantee any api stability. The implementation may
change in the future in backwards incompatible ways without any notice or
warning!