LoadIdentityClut: Make AMD fallback path handling more robust on Linux.
Once we know the DisplayCoreId is 'AMD', we know already it is an
AMD display engine, so once we know it is Linux, we can do our
thing instead of double-checking if it is really really really an
AMD gpu and screwing up in the process, because the naming scheme
of some AMD gpu's has changed again wrt. our matching. This led
to not detecting the Linux+AMD case on a AMD Vega with DCE-12
display engine, and wrong choice of identity lut type 1 instead of
the proper type 0 for Linux 4.17+ with AMD DisplayCore.

See this thread for the fallout:

https://psychtoolbox.discourse.group/t/dithering-issue-with-amd-graphic-card-linux-ubuntu/3760/35

Note: This error was undetected, because we never hit this fallback
in the past due to dumb luck. We would not hit it if MMIO low-level
tricks wouldn't be broken for Vega's DCE-12, which they unfortunately
are. Iow. more work is needed for Vega DCE-12 enablement if we want
to keep our bag of tricks working on those gpus.