PsychVulkanCore/OSX: Detect more broken macOS Metal timestamping.
If Metal delivers a zero timestamp signalling presentation failure for the
given present requesst, then MoltenVK maps that to .desiredPresentTime,
which covers up the failure! Make sure to make the failure explicit, by
comparing reported actualPresentTime against desiredPresentTime. If they
match, then that almost certainly means that Metal reported a zero timestamp,
ie. this was a present failure. In this case we report a zero timestamp, so our
caller knows things went badly.

In other news: I found out that timeout abort is not caused by feedback for
a presentId going missing. The presentation completion callback for that
presentId is called, but delayed beyond the timeout period and out of sequence,
e.g., after the callback for a *later* presentId, ie. one later in time. Our
fetched present history array does contain a record for the "timed out"
presentId eventually, just too late for us to do something meaningful with it.
In such cases, the present completion timestamp will be zero, reporting a
failed present.

So the rule that each present request causes one callback invocation and
corresponding present record log entry eventually, seems to hold. Failed
presents will have a zero timestamp, and may come out of order and with
large delay!