PsychPaidSupportAndServices: Allow to omit OrderId at users request.
Some users have trouble finding the OrderId. Add instructions on
how to find the OrderId.

If the user still can not find it, allow to instead (ab)use the
first 8 characters of the license key as OrderId.

This exposes part of the secret license key to the public, so
weakens the authentication, but then we are not the secret
service or such, so this is an acceptable risk-convenience tradeoff.

Also make input a bit more robust against trailing/leading
white-space.