Screen/Linux: A little bit of more fixup for Vega/DCE-12 identity passthrough.
If LoadIdentityClut() triggers low-level passthrough setup on AMD + DC,
and an identity lut is already loaded, then the gpu's regamma mode would
be "regamma bypass" - ie. an identity lut is detected, as we want it. But
as a side-effect, the hw would also report an all-zero hw lut in this
case, so we'd would also detect isZero = true, and as isZero reporting
takes precedence on isIdentity reporting, we'd report a zero-lut instead
of an (effective) identity lut, which leads our logic astray, trying to
PsychOSKDLoadIdentityLUT() a new hw identity lut - something not actually
supported on Vega/DCE-12 -- and we fail low-level hw identity setup,
triggering the fallback path and false reporting of failure, although
everything was already setup right -- ie. regamma bypass was active!

In practice, this caused the correct gpu behaviour, but confusing false
reporting of failure -- a false positive.

Fix this: If DisplayCore is in use and we detect regamma bypass, ie.
isIdentity == true, then force isZero = false, as an effective identity
lut is obviously not an all-zero lut. And all is good again, as verified
on a AMD Vega/DCE-12 with a Datapixx3.

[Work time spent on Linux + Vega/DCE-12 enabement so far: 17.5 hours]