PsychVulkan/OSX: Require at least macOS 10.15.4 for Vulkan backend.
10.15.4 is the minimum, as important Metal api is missing on earlier macOS
versions. Testing only happened on 10.15.7 Catalina final though, so that is
all we officially "support" for experimental use and testing.

In practice, this is broken/unreliable wrt. timing on AMD gpu's and an even
bigger disaster on Intel graphics chips, where reported timestamps are off
by a factor 1000x, apparently Apple engineering doesn't understand basic
multiplication or correctness testing -> This is definitely yet another Metal
bug.

Maybe it isn't as bad on macOS 11, who knows? On macOS 10 it is useless
for anything requiring even minimum reliability.