Add new PsychVulkanCore mex driver for Octave and Matlab on macOS.
Highly experimental. Requires the Khronos MoltenVK driver for Vulkan on
Metal support for macOS. Requires at least macOS 10.15.4, but was only
tested on macOS 10.15.7 Catalina final.

This uses Vulkan as a bridge between OpenGL and Metal, so effectively we
use Apple macOS Metal as display backend.

Testing on Catalina with AMD and Intel graphics shows that Metal on macOS
10 is as broken wrt. visual stimulation timing and timestamping as Apples
ancient OpenGL implementation - even more so for our use cases. So this
is only for the brave, masochistic or moronic, depending on how wants to word
it.

Hopefully Metal on macOS 11 would be less awful, but no testing was
performed on this due to lack of a safely usable independent test machine.

Work time spent so far: 162 hours, most of them quite depressing.