DrawFormattedText2: Fix UTF-8 string handling WRT Octave 6
With Octave 6, regexp expects strings to be truly UTF-8 encoded. This causes an "regexp: the input string is invalid UTF-8" error when calling DrawFormattedText2 with a vector that contains UTF character codes between 128 and 255, for example UTF-code 252 (umlaut-u). This can be easily reproduced with DrawFormattedText2Demo by inserting code 252 into the "unicodetext" vector near line 600.
As a workaround, we replace UTF codes >127 by zero rather than just replacing codes >255 by zero as it has been done so far.
This also works for Matlab.