CharAvail/GetChar: Adapt to Octave 6.1 cli on Catalina.
The behaviour of the trainwrecks has changed again.
Now on macOS + Octave in command line / non-GUI mode, we must not
use keystroke redirection, the opposite of before!

The entertainment never stops with the iToys.

ListenChar for octave cli mode is still unfixed, but my patience for this week
is exhausted. octave --gui seems to work fine though, so that's good enough.