PsychHID/OSX: Try to deal with 'Input monitoring' security restrictions.
On macOS 10.15 Catalina, a new api function IOHIDCheckAccess() was
introduced to the trainwrecks. It allows to check if input monitoring, as
needed for PsychHID HID input, is denied due to macOS security bullshit.

Now we try to dynamically bind to this function, and use it if it is supported.
If we detect security bs in our way, we abort PsychHID init at the earliest
possible moment, providing the users with troubleshooting instructions to
deal with the issue. Error abort implies closing onscreen windows, so the
user will not be left puzzled by a frozen window or such. We also request the
security approval dialog to aid the user a bit more.

However, the dialog will only pop up the first time we make this request,
because -- well nobody except the iToys knows why that would make any
sense. Successive requests will do nothing, and the user has to set up
approval manually.

-> This is another workaround for flawed security UX, courtesy of Apple,
and the user has to manually approve input monitoring in a tedious multi-
step procedure. Nothing we could do about this. But at least we now have
a way to diagnose the issue automatically instead of having to go on a
tedious diagnostic scavenger hunt each time the problem shows up for a
user.