Screen/Linux: Enable use of VRR mode with Vulkan display backend.
If VRR/FineGrained stimulation is requested and also use of the
Vulkan display backend, then select kPsychVRRSimple mode, ie.
map kPsychVRRAuto => kPsychVRRSimple. This way, VRR mode gets
enabled in the display driver (via our OpenGL/Mesa path), but
from there on, all flip scheduling works exactly as Vulkan
presents in fixed refresh rate mode would work. Ie. we call
PsychVulkanCore('Present', ...) with the tWhen target onset
timestamp and leave everything to PsychVulkanCore.

PsychVulkanCore will use the Vulkan timing extension to request
a Vulkan present at the specified onset time. From there on it
is completely up to the Vulkan ICD + operating system to do what
needs to be done for optimal VRR present.

This does require a Vulkan ICD with proper presentation timing
support for this, which is the difficult to-be-done part for
Linux. But this is all outside the scope of Psychtoolbox itself,
at least in this early phase of experimentation.