Python/hid.py: Use kbqeue default window handle 'None' instead of 0.
Python can't make use of the windowHandle parameter of PsychHID('KbQueueCreate') yet., as meaningful values would come from with Screen(), which isn't supported yet.
Safe choices atm. would be None or simply leaving the parameter out completely, so lets default to None.
Note that our docs say the default if omitted would be zero, which is sort-of correct. We mapped to 0 up to PTB 3.0.16, since PTB 3.0.17 something we map to a memory NULL pointer, which itself maps back to 0 for all practical purposes. This change was needed to accommodate potential future other windowing systems than Linux+X11.

Neither Matlab nor Python do support pointers in the scripting language afaik. On Matlab/Octave the parameter will take a pointer encoded in a uint32 (for 32-Bit systems) or uint64 (for 64-Bit systems), and so a 0 is indeed the same as a NULL pointer if the parameter is left out. On Python it takes a PyLong, which is converted into a pointer via Python PyEx api However, currently no user of this exists in PTB's Python extensions, and KbQueueCreate is the only meaningless user as long as we don't have Screen() for Python. Long story short, proper use of this parameter was only very lightly tested under Python. The strange thing is that different setups seem to be behaving differently wrt. accepting a 0 value, although the underlying C code is identical for operating systems and architectures, calling the same Python interpreter api's internally. Maybe some quirks of different Python versions / interpreters depending on OS and architecture? Acts of god? Who knows.

This will need more careful testing once more PTB functions would be ported to Python. For now the safe choice is to leave the parameter out, or set it to 'None'.

Hopefully fixes #723 but untested here.