Improve/Fix the procedural CreateProceduralSmoothedApertureSineGrating() shader.
The CreateProceduralSmoothedApertureSineGrating() procedural shader is implemented by the SineGratingSmoothedApertureShader fragment shader.

This shader smoothes the edge of a sine grating. By default the 'useAlpha' option is turned OFF and
smoothing uses the background color directly (alpha channel untouched). In this case no change will
be seen by this fix.

However when 'useAlpha' is turned on, the shader should use the edge modifier (Mod)on the alpha channel directly. Previously the fragment shader combined the global and local alpha ((rawAlpha * Mod) + Offset.a) causing an undesired effect (as I originally contributed this it was my bug), and the edge is not smoothed. To fix this we use the Modifier value directly (Offset.a * Mod), which allows the edge to blend properly with the underlying color while leaving the rest unaffected. When the background colour and offset are the same (as is normally the case, e.g. [0.5 0.5 0.5 1]), then useAlpha ON/OFF now looks the same. If the background colour is different to the offset, now with useAlpha=ON the edge properly smooths.

I added an optional value = 2 to 'Method' which copies what is done in the SmoothedDiscShader.frag (option to invert the smoothing mask). Previously it accepted 0 or 1, and so there will be no change unless you explicitly set this to 2.

In addition I added the .vscode folder to .gitignore, this folder gets generated if you use VS Code or VS Codium and shouldn't be tracked in git.

Authored-by: iandol <iandol@cog5>
Some rewording of the commit message by Mario Kleiner.