PsychPortAudio/Linux: Work around broken timing of ALSA pulseaudio plugin.
By default we do not allow use of Pulseaudio in parallel to PPA operation
at all, because this way (Pulseaudio server force-suspend) we get exclusive
access to the sound hardware for best possible timing.

But if the 'EngineTunables' override is used to allow to keep Pulseaudio
running in parallel to our operation, things are different. If Pulseaudio
is idle for our target sound device, everything is the same. But if it
is active, we may get redirected in our choice of default sound card to
the ALSA 'default' or 'pulse' device, which both map to the ALSA Pulseaudio
plugin. That plugin allows parallel sound i/o with PulseAudio, arbitrated
and mixed by PulseAudio. The problem is that the plugin does not return any
audio timestamps - all timestamps are constant zero!

So far we simply returned bogus timestamps and screwed up totally if
timed audio onset/offset/scheduling was requested. We can do a bit better:

If we detect all-zero timestamps from Pulseaudio, we replace them by
current system time aka GetSecs time. This gives somewhat wrong/inaccurate
but at least consistent timestamps and allows our client code to run with
an accuracy that is at least good enough for basic auditory feedback,
instead of going bong-shit. This is all consistent with the QoS promised
in PPA 'Open' for 'reqlatencyclass' == 0, e.g., as used by simple demos
like BasicSoundOutputDemo.m or the Snd() and Beeper() functions. If user
code requested reqlatencyclass > 0, ie. low latency, high timing precision,
then we print a warning each time PsychPortAudio('Stop') is called if
timing and timestamping was broken during a auditory stimulation run, so
there aren't any surprises or sad faces in the end.