PsychOpenHMDVR: Fixup for recent Linux kernels and distros.
Turns out that at least on Ubuntu 20.04-LTS with modern Linux kernels,
which are HMD aware, for at least the Oculus Rift CV1, our old approach
with PsychOpenHMDVR() does not work well anymore. The Oculus Rift CV1
and later Oculus models go into a power-saving mode at the end of a
session when our driver closes the USB connection to the HMD device and
no more USB keep-alive packets are sent anymore.

When entering power-save, the HMD's video output, e.g., HDMI, is
hot-unplugged from the system. On hot-replug (ie. when a new session
is started, our driver connects vie USB, and the Rift fully powers up),
the Rift video display does not fully turn on anymore. It seems that
the panel stays powered down due to some "stuck" DPMS power setting.
This may be a quirk of the specific Rift CV1 HMD, or it may be
because the HMD is marked as non-desktop display by Linux 4.16+, and
either the kms driver or the X-Server somehow gets confused by the
hot-unplug without proper DPMS handshaking, or some other problem of
some other component. In any case, this problem was observed on the
Rift CV1 with Ubuntu 20.04-LTS, standard distro kernel 5.8, X-Server
20.0.9. The problem does not happen as often and reproducible on a
Linux 5.13 kernel, but sometimes it still happens.

Typically the Rift works for exactly one session after reboot, then
not anymore.

I found the following workaround implemented in this commit:

During the device open sequence / driver session startup:

1. Explicitely turn off the RandR video output assigned to the HMD.
2. Wait a second.
3. Turn on the HMD RandR output again, at the correct native HMD panel
   mode and refresh rate. Also label the output as standard desktop
   display, overriding the Linux kernels classification/labelling as
   non-desktop special purpose display, which is applied to known HMD's
   since Linux 4.16.
4. Wait another second or two.

What this output-off --> output-on sequence implicitely does is, that
it first sends a DPMS standby signal to the HMD, then a DPMS on signal,
enforcing a toggle of the panels DPMS state. This gets the display unstuck.
We could also just use calls to force DPMS standby -> DPMS on, without
the need for a full modeset, but as DPMS acts globally on all X-Screens
and outputs, this causes flicker and short display blanking on all active
displays, not just the HMD, so it is more intrusive / annoying.

Note we use xrandr system() callouts for this, as our own builtin
Screen() modesetting code doesn't deal with turned off outputs yet,
and just updating Screen() for this seems a bit too much overkill
at the moment.

Another note: In order to make multi-X-Screen configurations work
at all, which is needed with our current Linux VR drivers for good
timing and performance, the xorg.conf Monitor section for the HMD's
ZaphodHeads HMD output needs an explicit...

Option        "Enable" "on"

...to force the X-Server to enable the output despite it being
marked as non-desktop by the Linux kernel. Otherwise the server
will not find any connected outputs, is left with an X-Screen
without outputs and then deletes that X-Screen, and we are left
with a single-X-Screen setup that is unusable for VR HMD's.

XOrgConfCreator can not auto-generate such files, so i updated the
"help PsychOpenHMDVR" text with setup instructions. We also ship
a new sample xorg.conf file demonstrating this under:

Psychtoolbox/PsychHardware/LinuxX11ExampleXorgConfs/xorg.conf_DualXScreen_OculusRift_amdgpu.conf

Another note: If one wants to use the monado-service VR compositor
for OpenXR on Linux, one needs a bug-fixed monado implementation and
Linux kernel 5.13 or later - or at least it is verifed to work with
5.13 and to not work with 5.8. Current monado master will only work
on single-X-Screen setups which are unsuitable for parallel use with
our old VR drivers.

Another note: Getting at least the Rift working under these circumstances
on a single X-Screen setup proved to be painful to the point of being
unworkable. The Rift's power-saving collaborates with the Linux 4.16
non-desktop display labelling to create a perfect shit-show of pain.
Not impossible to get working, but painful enough that i won't bother,
so i doubt any normal users will. Of course this doesn't matter much,
as the single-X-Screen config is only useful for the most basic testing,
if one is to lazy to do a logout -> Switch to dual-x-screen -> login
transition. Performance is by design so bad, it is not useful for actual
VR experiments, and setting up a dual-X-Screen setup only takes a minute,
so there...

-> This is all quite a bit of a hack, but given that i am working on a new
   VR driver anyway, which intends to supersede all our existing drivers,
   and does not suffer any of these problems, i guess patching this up a
   tiny bit more should do for the time being to bridge the time-gap between
   now and the completion of the new driver in a few months.