PsychOpenHMDVRCore: Add head speed/acceleration reporting - disabled for now.
Add queries and reporting of head linear/angular velocity and
acceleration, based on future OpenHMD api. This is not yet
merged into master or part of an official OpenHMD release. Also
even the experimental tracking branches only support this for the
Rift DK2/CV1, and only incomplete. The library spams the Octave/
Matlab console with error messages about unsupported queries on
both the current public OpenHMD releases and on the experimental
one for non-Rift, which is unacceptable. Therefore we leave this
tested implementation disabled for now.

It only gets enabled if the OpenHMD version is greater than a so
far unknown future version - iow. not for any shipping version atm.

PsychOpenHMDVR() contains corresponding disabled code for now.