PsychOpenHMDVR(): Fixup ipd handling, make consistent with OpenHMD behaviour.
Rework the math of how ipd (==interpupillar distance) is used for
rendering. This makes our modelView matrices consistent with the
ones that libOpenHMD would calculate, which is more reassuring.

Testing shows no ill effects in either 3D or 2D stereo, or 2D mono
rendering scenarios like VRHMDDemo, ImagingStereoDemo for 2D rendering,
or VRHMDDemo1 and SuperShapeDemo and MorphDemo for 3D rendering.

-> This commit establishes our new baseline for VR rendering on
   Oculus Rift CV1 and other HMDs other than the legacy Rift DK1
   and Rift DK2, ie. where we can not use the legacy PsychOculusVR
   driver via the Oculus VR 0.5 SDK.

-> Ready for the actual work.