Screen: Suppress broken pageflip warning if external display method is in use.
Also if the window is not a fullscreen window, as failure is expected then.

This checking code only ever executes on Linux + AMD if the user foolishly
disables OpenML timing and timestamping, ie. never under normal circumstances,
or on macOS with PTB kernel driver loaded on a AMD or Intel gpu.

In practice, therefore only the macOS + kernel driver case matters, and
only for standard OpenGL display backend via CGL. Failure with the Cocoa
backend is both expected and also irrelevant if the external display method,
ie. the Vulkan + MoltenVK == Metal backend is used. Therefore we suppress
this message with external (Metal) display backend active, as that backend
forces use of Cocoa and thereby always implies this failure - iow. produces
a false alert.