PsychOculusVRCore1: Shutdown runtime on last HMD close.
Current Oculus VR runtime shows a bright flashing screen if the last HMD has
closed, but the application has not yet called ovr_Shutdown() to disconnect from
the VR compositor. Iow. we get bright annoying white blinking until a "clear all" or
similar. This is annoying, so now we ovr_Shutdown() as soon as the last HMD device
is closed.