PsychOculusVR1: Fix white-flash at end of session.
We need to perform a wait for  a second before closing the PsychOculusVRCore1
driver, i.e. before HMD close and releasing the connection to the VR runtime/server,
otherwise the compositor flashes some all-white image in some weird animation for
about 0.5 seconds. This did not happen in earlier times, so maybe a bug or behavioural change in the Oculus VR runtime? Anyway, easily enough worked around.

Also update copyright dates for 2021 when we are here, and some code comments.