VRInputStuffTest: Use Enter button for advance or exit.
The Enter button (little Hamburger on the left touch controller and on the
X-Box controller, as well as central round button on Oculus remote) is a
better choice for stepping through the test or ending it, than the previously
used Home button (The Oculus symbol on the right touch controller, and
Oculus remote), because while the Home/Oculus button works fine on Linux,
on Windows it will switch VR focus from Psychtoolbox to the Oculus Home app,
preventing any detection of the button press, or progress for PTB. Iow. the
Oculus button should be generally avoided during a PTB session.

The Back button is still an additional valid choice, as it maps to the Back-Arrow
key on the Oculus remote, and the little Menu key on the XBox controller.