PsychVulkan/Linux: Improve workaround for NVidia proprietary Vulkan driver.
Turns out that direct display mode works with the NVidia proprietary
blob if the kernel boot option nvidia.drm_modeset=1 is not specified,
which apparently triggers some bugs in some nvidia blobs, although those
are supposedly fixed in the latest NVidia drivers. Luckily this option
is not the default, so things will just work on setups with default
installation.

However, it turned out that our old method of reenabling the RandR
output via xrandr --auto also restored the viewport of the output
to X-Screen offset 0,0 ie. the output was always reactivated at the
end of a session in mirror/clone mode, which is confusing to users
and also later on to the window rectangle matching logic of PsychVulkan
if another Vulkan session is run. The user was forced to manually
restore the displaz arrangement in screen space via xrandr or GUI
setting tools -- Suboptimal.

Now we remember the old location of the viewport of the to-be-turned-off
output on which we want to do direct display presentation under NVidia
Vulkan. At Vulkan window close time, we specify the old x,y offset of
the viewport to the xrandr --auto command via the --set option, so
display arrangement is restored to the state before PTB session startup,
and all is good. Or so we hope...

One catch: This will go wrong, and still switch to mirror mode, if the
to-be-used output in a multi-display setup is the output located at
X-Screen x,y offset (0,0) ie. defines the desktop origin. This because
the window manager will detect that the output got disconnected and then
it will shift remaining outputs into the (0,0) position to close the
hole that was left by the disabled output. At end of session, the output
gets properly reenabled at the origin, but the other unused output(s) will
stay where the window manager put them at session startup / output disable,
ie. they will now mirror the former Vulkan output!

However, this should not be a problem on single-display per X-Screen
setups, where no shifting around can happen, or on multi-display setups
where the subject / PTB visual stimulation outputs will usually be not
located in the X-Screen origin (0,0) in the first place, as that place
is usually reserved for the experimenter GUI display with Octave or
Matlab.

Note that this commit is untested on an actual NVidia gpu with the
actual NVidia driver. I only simulated the setup on a dual-display
AMD setup with amdgpu driver, by forcing needsNvidiaWa = 1. Here things
work fine now under both Mesa radv and AMD amdvlk, so i expect them to
work (at least better than before) on an actual NVidia proprietary setup.

-> Actual test pending.