Snd(): Update help text a bit, add "permanent" PPA sharing.
The new Snd('Open', pahandle, 1); call - with the optional '1' as
2nd argument, will disable Snd('Close'). Iow. Snd('Close') will
only stop sound playback like Snd('Quiet'), but not close Snd() at
all, not even detach from the shared pahandle. This can be overriden
via Snd('Close', 1) instead of the old Snd('Close').

Why? Because of Beeper(), which always calls Snd('Close') after each
beep, which makes sense for the simple Beeper() use pattern, but will
detach Snd() from a shared pahandle if such a handle is shared with
PsychPortAudio(), so in the shared case one can only emit one beep
tone on such a shared device. This is unexpected by users. Worse,
Beeper() is used by the default auditory feedback implementation for
Eyelink toolbox gazetracker calibration, so the calibration was only
able to emit one beep tone, and got stuck after that due to the
unexpectedly closed Snd() pahandle, if PsychPortAudio was used in
parallel during an Eyelink session.

Now user code can Snd('Open', pahandle', 1) such a handle instead,
thereby turning Snd('Close') into a no-op, thereby prevent Beeper()
from being a one-hit wonder, thereby keeping Eyelink auditory feedback
from screwing up during tracker calibration.

The existence of Snd() and use by Beeper() and Eyelink is unfortunate,
but we have to keep this hackery around for backwards compatibility and
hope that the "help Snd" text will scare users into doing the right
thing and not using Snd()/Beeper() etc. for anything non-trivial.

See https://psychtoolbox.discourse.group/t/audio-device-issues-jack-server-in-ubuntu/3963
for issues caused by the old implementation in conjunction with Eyelink,
in this case on Linux, but the same problem would apply on MS-Windows,
and would cause massive timing issues on macOS even if sharing would appear
to work, but the failure on macOS would be more subtle and "silent", and
so even worse if the user would not note the broken timing...

Effort for this work: 80 minutes so far.